+++ inst/IAOReader/server/modals.R
@@ -20,23 +20,46 @@ modals <- function(input, output, session) {
 
     # K parameter info ---------------------------------------------------------
     observeEvent(input[["k_param_info"]], {
-        modal <- k_param_info_modal()
+        # No return, link to measure info modal with return.
+        modal <- k_param_info_modal(NULL, "measure_info_return")
         showModal(modal)
     })
 
     observeEvent(input[["k_param_info_return"]], {
-        modal <- k_param_info_modal("measure_info")
+        # Return to measure info modal without return, link to measure info
+        # modal with return.
+        modal <- k_param_info_modal("measure_info", "measure_info_return")
+        showModal(modal)
+    })
+
+    observeEvent(input[["k_param_info_return_plot"]], {
+        # Return and link to measure info modal which returns to plot info
+        # modal.
+        modal <- k_param_info_modal(
+            "measure_info_return_plot",
+            "measure_info_return_plot"
+        )
         showModal(modal)
     })
 
     # Measure info -------------------------------------------------------------
     observeEvent(input[["measure_info"]], {
-        modal <- measure_info_modal()
+        # No return, link to k param info modal with return.
+        modal <- measure_info_modal(NULL, "k_param_info_return")
         showModal(modal)
     })
 
     observeEvent(input[["measure_info_return"]], {
-        modal <- measure_info_modal("k_param_info")
+        # Return to k param info modal without return, link to k param info
+        # modal with return.
+        modal <- measure_info_modal("k_param_info", "k_param_info_return")
+        showModal(modal)
+    })
+
+    observeEvent(input[["measure_info_return_plot"]], {
+        # Return to plot info modal, link to k param info modal which returns to
+        # this modal.
+        modal <- measure_info_modal("plot_info", "k_param_info_return_plot")
         showModal(modal)
     })